requirements.txt
langgraph
langchain-groq
langchain-community
tavily-python
yfinance
python-dotenv
yagmail
supabase
markdown
matplotlib